Implement rtimers for ADuCRF101

+++ cpu/arm/aducrf101/aducrf101-contiki.h
@@ -45,8 +45,8 @@
 typedef uint32_t clock_time_t;
 typedef uint16_t uip_stats_t;
 
-typedef unsigned long rtimer_clock_t;
-#define RTIMER_CLOCK_LT(a, b)     ((signed long)((a) - (b)) < 0)
+typedef uint32_t rtimer_clock_t;
+#define RTIMER_CLOCK_LT(a, b)  ((int32_t)((a) - (b)) < 0)
 rtimer_clock_t rtimer_arch_now(void);
 
 #endif
